refactor(backend): split lizyml_ui_schema into constants/metrics modules (M-1)
Split lizyml_ui_schema.py (714 lines) into 3 focused modules:
- lizyml_constants.py: param keys, additional params helper
- lizyml_metrics.py: metric evaluation/direction with caching
- lizyml_ui_schema.py: build_ui_schema() only

All import paths maintained via re-export. 764 tests pass.
