+++ src/lizystudio/backends/lizyml_ui_schema.py
@@ -6,187 +6,15 @@ Contains only static data structures — no ML logic.
 
 from __future__ import annotations
 
-import threading
 from typing import Any
 
-# Keys already covered by parameter_hints or search_space_catalog.
-# additional_params must exclude these to avoid UI duplication.
-_KNOWN_PARAM_KEYS: frozenset[str] = frozenset(
-    {
-        "objective",
-        "metric",
-        "n_estimators",
-        "learning_rate",
-        "max_depth",
-        "max_bin",
-        "feature_fraction",
-        "bagging_fraction",
-        "bagging_freq",
-        "lambda_l1",
-        "lambda_l2",
-        "first_metric_only",
-        "verbose",
-        "num_threads",  # Excluded from UI intentionally (server-managed)
-        "num_leaves",
-        # Smart params (in search_space_catalog)
-        "auto_num_leaves",
-        "num_leaves_ratio",
-        "min_data_in_leaf_ratio",
-        "min_data_in_bin_ratio",
-        "feature_weights",
-        "balanced",
-    }
+from lizystudio.backends.lizyml_constants import _build_additional_params
+from lizystudio.backends.lizyml_metrics import (
+    get_eval_metrics_by_task as get_eval_metrics_by_task,
 )
-
-# LightGBM parameters available as additional params (beyond hints/catalog).
-_LGBM_ADDITIONAL_PARAMS: list[str] = sorted(
-    [
-        "min_child_weight",
-        "min_child_samples",
-        "subsample",
-        "colsample_bytree",
-        "reg_alpha",
-        "reg_lambda",
-        "max_cat_threshold",
-        "cat_smooth",
-        "cat_l2",
-        "extra_trees",
-        "path_smooth",
-        "min_gain_to_split",
-        "min_data_in_leaf",
-        "min_data_in_bin",
-        "max_cat_to_onehot",
-        "top_k",
-        "min_sum_hessian_in_leaf",
-        "linear_tree",
-        "feature_pre_filter",
-        "force_col_wise",
-        "force_row_wise",
-        "histogram_pool_size",
-        "is_unbalance",
-        "scale_pos_weight",
-        "sigmoid",
-        "boost_from_average",
-        "bin_construct_sample_cnt",
-        "data_sample_strategy",
-        "interaction_constraints",
-    ]
+from lizystudio.backends.lizyml_metrics import (
+    get_metric_directions,
 )
-
-
-def _build_additional_params() -> list[str]:
-    """Return LightGBM params not in parameter_hints or search_space_catalog."""
-    return [p for p in _LGBM_ADDITIONAL_PARAMS if p not in _KNOWN_PARAM_KEYS]
-
-
-# --- Eval metrics registry lookup (cached) ---
-
-_eval_metrics_cache: dict[str, list[str]] | None = None
-_eval_metrics_lock: threading.Lock = threading.Lock()
-
-# Preferred metric per task — shown first in UI (Widget conformance).
-_PREFERRED_METRIC: dict[str, str] = {
-    "binary": "auc",
-    "regression": "rmse",
-    "multiclass": "auc",
-}
-
-
-def _sort_with_preferred(metrics: list[str], task: str) -> list[str]:
-    """Sort metrics alphabetically but place the preferred metric first."""
-    preferred = _PREFERRED_METRIC.get(task)
-    ordered = sorted(metrics)
-    if preferred and preferred in ordered:
-        ordered = [preferred, *(m for m in ordered if m != preferred)]
-    return ordered
-
-
-def get_eval_metrics_by_task() -> dict[str, list[str]]:
-    """Query LizyML's metric registry for available evaluation metrics per task."""
-    global _eval_metrics_cache  # noqa: PLW0603
-
-    if _eval_metrics_cache is not None:
-        return _eval_metrics_cache
-    with _eval_metrics_lock:
-        if _eval_metrics_cache is not None:
-            return _eval_metrics_cache
-        metrics: dict[str, list[str]]
-        try:
-            from lizyml.metrics.registry import _TASK_METRICS
-
-            metrics = {
-                task: _sort_with_preferred(list(ms), task)
-                for task, ms in _TASK_METRICS.items()
-            }
-        except (ImportError, AttributeError, TypeError):
-            # Fallback for older LizyML versions
-            metrics = {
-                "regression": _sort_with_preferred(
-                    ["mae", "mape", "rmse", "huber", "r2", "rmsle"], "regression"
-                ),
-                "binary": _sort_with_preferred(
-                    [
-                        "auc",
-                        "logloss",
-                        "auc_pr",
-                        "f1",
-                        "accuracy",
-                        "brier",
-                        "ece",
-                        "precision_at_k",
-                    ],
-                    "binary",
-                ),
-                "multiclass": _sort_with_preferred(
-                    ["logloss", "f1", "accuracy", "auc", "auc_pr", "brier"],
-                    "multiclass",
-                ),
-            }
-        _eval_metrics_cache = metrics
-        return metrics
-
-
-# --- Metric direction lookup (cached) ---
-
-_metric_direction_cache: dict[str, dict[str, str]] | None = None
-_metric_direction_lock: threading.Lock = threading.Lock()
-
-
-def get_metric_directions() -> dict[str, dict[str, str]]:
-    """Return {task: {metric: 'minimize'|'maximize'}} from LizyML registry."""
-    global _metric_direction_cache  # noqa: PLW0603
-
-    if _metric_direction_cache is not None:
-        return _metric_direction_cache
-    with _metric_direction_lock:
-        if _metric_direction_cache is not None:
-            return _metric_direction_cache
-        result: dict[str, dict[str, str]] = {}
-        metrics_by_task = get_eval_metrics_by_task()
-        try:
-            from lizyml.metrics.registry import get_metric
-
-            for task, metric_names in metrics_by_task.items():
-                task_dirs: dict[str, str] = {}
-                for name in metric_names:
-                    try:
-                        m = get_metric(name)
-                        task_dirs[name] = (
-                            "maximize" if m.greater_is_better else "minimize"
-                        )
-                    except Exception:  # noqa: BLE001
-                        task_dirs[name] = "minimize"
-                result[task] = task_dirs
-        except (ImportError, AttributeError):
-            # Fallback
-            maximize = {"auc", "auc_pr", "r2", "accuracy", "f1", "auc_mu"}
-            for task, metric_names in metrics_by_task.items():
-                result[task] = {
-                    m: "maximize" if m in maximize else "minimize" for m in metric_names
-                }
-        _metric_direction_cache = result
-        return result
-
 
 # --- UI schema builder ---
